Port to CMake and make some stylistic changes
Removed the autotools files and the msvc8 project files, replacing them
with CMake. To follow similar projects I've moved the files around so
the tests are in the ~/tests/ directory and the library itself is at the
root of the project (where the CMakeLists.txt file is as well).

I find it easier to have all-lowercase filenames and have C++ headers
end with .hpp rather than .h -- following other projects such as Boost.
So I've made this particular change as well.